Консольный файловый менеджер Выполнены задания к занятию №8

+++ file_manager.py
@@ -5,44 +5,34 @@ from menu_functions import bank_account
 import sys
 from menu_functions import isdir
 from menu_functions import isfile
+from menu_functions import main_menu_list
+
 
 while True:
 
-    print('='*10)
-    print('1 - создать папку')
-    print('2 - удалить (файл/папку)')
-    print('3 - копировать (файл/папку)')
-    print('4 - просмотр содержимого рабочей директории;')
-    print('5 - посмотреть только папки')
-    print('6 - посмотреть только файлы')
-    print('7 - просмотр информации об операционной системе')
-    print('8 - создатель программы')
-    print('9 - играть в викторину')
-    print('10 - мой банковский счет')
-    print('11 - выход')
-    print('=' * 10)
+    main_menu_list()
 
     menu_choice = input('Выберите пункт меню: ')
     if menu_choice == '1':
         dir_name = input('введите название папки: ')
-        if not os.path.exists(dir_name):
+        try:
             os.mkdir(dir_name)
+        except:
+            print('папка с таким именем уже существует')
 
     elif menu_choice == '2':
         remove_name = input('введите название папки/файла: ')
-        if os.path.exists(remove_name):
-            if os.path.isfile(remove_name):
-                os.remove(remove_name)
-            else:
-                shutil.rmtree(remove_name)
+        try:
+            os.remove(remove_name) if os.path.isfile(remove_name) else shutil.rmtree(remove_name)
+        except:
+            print('папки/файла с таким именем не найдено')
 
     elif menu_choice == '3':
         copy_name = input('Введите название папки/файла')
-        if os.path.exists(copy_name):
-            if os.path.isfile(copy_name):
-                shutil.copy(copy_name, f'copy_{copy_name}')
-            else:
-                shutil.copytree(copy_name, f'copy_{copy_name}')
+        try:
+            shutil.copy(copy_name, f'copy_{copy_name}') if os.path.isfile(copy_name) else shutil.copytree(copy_name, f'copy_{copy_name}')
+        except:
+            print('папки/файла с таким именем не найдено')
 
     elif menu_choice == '4':
         print(os.listdir())
@@ -53,6 +43,7 @@ while True:
                 dirs = isdir()
                 with open('listdir.txt', 'w') as f:
                     f.write(f'files: {files}\ndirs: {dirs}')
+                break
             elif save_dir == 'нет':
                 break
             else:
